Another migration fix

+++ src/kz/db/migrations.cpp
@@ -106,7 +106,7 @@ void KZDatabaseService::CheckMigrations(std::vector<ISQLQuery *> queries)
 	for (u32 i = 0; i < current; i++)
 	{
 		result->FetchRow();
-		u32 currentCRC = result->GetInt(1);
+		u32 currentCRC = result->GetInt64(1);
 		std::string migrationQuery;
 
 		switch (KZDatabaseService::GetDatabaseType())
